# docs/swagger/swagger.yaml
basePath: /v1
definitions:
...
  models.Income:
    properties:
      id:
        type: string
      lastUpdate:
        type: string
      netIncome:
        type: string
      note:
        type: string
      submitDate:
        type: string
      totalIncome:
        type: string
      userId:
        type: string
      vat:
        type: string
      wht:
        type: string
    type: object
  models.IncomeReq:
    properties:
      note:
        type: string
      totalIncome:
        type: string
    type: object
  models.IncomeStatus:
    properties:
      status:
        type: string
      submitDate:
        type: string
      user:
        $ref: '#/definitions/models.User'
        type: object
    type: object
...
  models.Site:
    properties:
      color:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
  models.User:
    properties:
      bankAccountName:
        type: string
      bankAccountNumber:
        type: string
      corporateName:
        type: string
      create:
        type: string
      email:
        type: string
      firstName:
        type: string
      id:
        type: string
      imageProfile:
        type: string
      lastName:
        type: string
      lastUpdate:
        type: string
      project:
        type: string
      role:
        type: string
      site:
        $ref: '#/definitions/models.Site'
        type: object
      siteId:
        type: string
      slackAccount:
        type: string
      thaiCitizenId:
        type: string
      transcript:
        type: string
      vat:
        type: string
    type: object
  utils.HTTPError:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
host: http://worklog-dev.odds.team/api
info:
  contact: {}
  description: This is a sample server odds-worklog server.
  license: {}
  title: Odds-Worklog Example API
  version: "1.0"
paths:
...
  /incomes:
    post:
      consumes:
      - application/json
      description: Add Income
      parameters:
      - description: ' '
        in: body
        name: income
        required: true
        schema:
          $ref: '#/definitions/models.IncomeReq'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Income'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPError'
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.HTTPError'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPError'
            type: object
      summary: Add Income
      tags:
      - incomes
  /incomes/{id}:
    put:
      consumes:
      - application/json
      description: Update Income
      parameters:
      - description: ' '
        in: body
        name: income
        required: true
        schema:
          $ref: '#/definitions/models.IncomeReq'
          type: object
      - description: Income ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Income'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPError'
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.HTTPError'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPError'
            type: object
      summary: Update Income
      tags:
      - incomes
  /incomes/current-month/{id}:
    get:
      consumes:
      - application/json
      description: Get Income Of Current Month By User Id
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Income'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPError'
            type: object
      summary: Get Income Of Current Month By User Id
      tags:
      - incomes
...
  /incomes/export/pdf:
    get:
      consumes:
      - application/json
      description: Get Export to Pdf file.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPError'
            type: object
      summary: Get Export Pdf
      tags:
      - incomes
  /incomes/status/corporate:
    get:
      consumes:
      - application/json
      description: Get Income Status List
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.IncomeStatus'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPError'
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.HTTPError'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPError'
            type: object
      summary: Get Corporate Income Status List
      tags:
      - incomes
  /incomes/status/individual:
    get:
      consumes:
      - application/json
      description: Get Individual Income Status List
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.IncomeStatus'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPError'
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.HTTPError'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPError'
            type: object
      summary: Get Individual Income Status List
      tags:
      - incomes
...
swagger: "2.0"